backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
chromadb==0.4.15
# Logics-Parsing: Use local installation at /Users/zhangcz/ws/python/src/github.com/alibaba/Logics-Parsing
# Required dependencies: modelscope, huggingface_hub, torch, transformers
langchain==0.1.0
crewai==0.28.8
langchain-community==0.0.29
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
click==8.1.7
python-docx==1.1.0